requirements.txt
mediapipe
numpy
opencv-python
tensorflow
cvzone
streamlit
streamlit_webrtc
av
PIL
keras.models